add fixed_length_segment
this consolidates the handling of splitting the input stream after knowing
the length of a sub-segment, and then passing the known end of that
sub-segment forward after parsing it.  we do this so that failure to consume
all of a sub-segment doesn't put the next segment starting at the wrong spot.