add timeout to dd and associated error handling
now if the SSH command(s) hang for some reason, we timeout attempting
to wait for them and will refresh our view of xochitl process ID and
frame buffer offset.  this (in theory) allows us to keep going when a
single `dd` command hangs, which i have seen happen a few times.